requirements.txt
regex
discord.py
deepl